+++ custom_components/wyoming_dotstts/__init__.py
@@ -22,7 +22,13 @@ from .const import CONF_API_TOKEN, CONF_HOST, CONF_PORT, DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
 
-PLATFORMS = [Platform.BINARY_SENSOR, Platform.NUMBER, Platform.SENSOR]
+PLATFORMS = [
+    Platform.BINARY_SENSOR,
+    Platform.NUMBER,
+    Platform.SELECT,
+    Platform.SENSOR,
+    Platform.SWITCH,
+]
 SCAN_INTERVAL = timedelta(seconds=60)
 REQUEST_TIMEOUT = aiohttp.ClientTimeout(total=15)